Move crypto code to the core module
