+++ packages/core/src/login/oidc/redirectHandler/IRedirectHandler.ts
@@ -40,3 +40,28 @@ type IRedirectHandler = IHandleable<
   RedirectResult
 >;
 export default IRedirectHandler;
+
+/**
+ * Extract a WebID from an ID token payload based on https://github.com/solid/webid-oidc-spec.
+ * Note that this does not yet implement the user endpoint lookup, and only checks
+ * for webid or IRI-like sub claims.
+ *
+ * @param idToken the payload of the ID token from which the WebID can be extracted.
+ * @returns a WebID extracted from the ID token.
+ * @internal
+ */
+export async function getWebidFromTokenPayload(
+  idTokenClaims: Record<string, string>
+): Promise<string> {
+  if (typeof idTokenClaims.webid === "string") {
+    return idTokenClaims.webid;
+  }
+  try {
+    const webid = new URL(idTokenClaims.sub);
+    return webid.href;
+  } catch (e) {
+    throw new Error(
+      `The ID token has no 'webid' claim, and its 'sub' claim of [${idTokenClaims.sub}] is invalid as a URL - error [${e}].`
+    );
+  }
+}
