+++ packages/oidc/package.json
@@ -33,13 +33,9 @@
   "dependencies": {
     "@inrupt/solid-client-authn-core": "^1.8.1",
     "@types/form-urlencoded": "^2.0.1",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/node-jose": "^1.1.5",
     "@types/uuid": "^8.3.0",
     "form-urlencoded": "~6.0.3",
     "jose": "^2.0.4",
-    "jsonwebtoken": "^8.5.1",
-    "node-jose": "^2.0.0",
     "oidc-client": "^1.11.3",
     "uuid": "^8.3.1"
   },
